Fixing warning

+++ Examples/StackExample/StackExample/ChildTestViewController.m
@@ -32,7 +32,7 @@
 - (IBAction)test:(id)sender
 {
     [self.stackedNavigationController pushViewController:[ChildTestViewController new] onTopOf:self animated:YES];
-	self.stackedNavigationController.tabBarItem.badgeValue = [NSString stringWithFormat:@"%d", self.stackedNavigationController.viewControllers.count];
+	self.stackedNavigationController.tabBarItem.badgeValue = [NSString stringWithFormat:@"%lu", (unsigned long)self.stackedNavigationController.viewControllers.count];
 }
 
 - (SPStackedNavigationPageSize)stackedNavigationPageSize;
